tests: Move MessageVisibilityTest to test_message_fetch.py.
